libragephoto 0.3.0 release
- update README.md
- update documentation
- ragephoto-gtkviewer: fix error uint8_t -> int32_t, NULL -> nullptr
- ragephoto-qtviewer: fix error uint8_t -> int32_t
- RagePhotoA: fix documentation, fix error()
- RagePhotoC: add brief

+++ examples/ragephoto-qtviewer/src/main.cpp
@@ -38,7 +38,7 @@ bool readPhotoFile(const QString &filename, QMainWindow *mainWindow, QLabel *pho
         RagePhoto ragePhoto;
         const bool loaded = ragePhoto.load(fileData.data(), static_cast<size_t>(fileData.size()));
         if (!loaded) {
-            const uint8_t error = ragePhoto.error();
+            const int32_t error = ragePhoto.error();
             if (error <= RagePhoto::PhotoReadError) {
                 QMessageBox::warning(mainWindow, "Open Photo", "Failed to read photo: " + filename);
                 return false;
